혜진

+++ cheatingday/src/main/java/com/icia/cheatingday/manager/dao/ManagerDao.java
@@ -56,7 +56,7 @@ public class ManagerDao {
 	public String findMusernameByMnum(long mNum) {
 		return tpl.selectOne("managerMapper.findMusernameByMnum", mNum);
 	}
-	public String findMirumeByMnum(int mNum) {
+	public String findMirumeByMnum(long mNum) {
 		return tpl.selectOne("managerMapper.findMirumeByMnum", mNum);
 	}
